Moves scoreboard to begining of flow, and styles it
