+++ src/pages/index.tsx
@@ -11,7 +11,7 @@ import { MovingPaddings } from "../components/stage/MovingPaddings";
 import { RememberNumber } from "../components/stage/RememberNumber";
 import { Ferdig } from "../components/stage/Ferdig";
 import { AreYouSure } from "../components/stage/AreYouSure";
-import {Resultatliste} from "../components/stage/Resultatliste";
+import {Toppliste} from "../components/stage/Toppliste";
 
 const Home: NextPage = () => {
   const context = useContext(AppContext);
@@ -20,34 +20,34 @@ const Home: NextPage = () => {
 
   switch (context?.page) {
     case 1:
-      stepComponent = <TakkForPameldelsen />;
+      stepComponent = <Welcome/>;
       break;
     case 2:
-      stepComponent = <AndreSmarteUtivklere />;
+      stepComponent = <TakkForPameldelsen />;
       break;
     case 3:
-      stepComponent = <FinnWilly />;
+      stepComponent = <AndreSmarteUtivklere />;
       break;
     case 4:
-      stepComponent = <ForklarHvorfor />;
+      stepComponent = <FinnWilly />;
       break;
     case 5:
-      stepComponent = <MovingPaddings />;
+      stepComponent = <ForklarHvorfor />;
       break;
     case 6:
-      stepComponent = <RememberNumber />;
+      stepComponent = <MovingPaddings />;
       break;
     case 7:
-      stepComponent = <AreYouSure />
+      stepComponent = <RememberNumber />;
       break;
     case 8:
-      stepComponent = <Ferdig />;
+      stepComponent = <AreYouSure />
       break;
     case 9:
-      stepComponent = <Resultatliste />;
+      stepComponent = <Ferdig />;
       break;
     default:
-      stepComponent = <Welcome />;
+      stepComponent = <Toppliste />;
       break;
   }
 
@@ -55,8 +55,7 @@ const Home: NextPage = () => {
     context.setTime(0);
   }
 
-  // "context?.page < 8" is only relevant during testing
-  if (context && context?.page > 0 && context?.page < 8 && !context.timerIsRunning) {
+  if (context && context?.page > 1 && context?.page < 9 && !context.timerIsRunning) {
     context.startTimer();
   }
